hdmitx: Fix parsing of HDR10plus and Dovi VSVDBs [1/1]
PD#SWPL-17786

Problem:
No dolby vision on latest Vizio TVs providing an EDID containing
both dolby vision and HDR10plus VSVDBs in the EDID.

Solution:
If an EDID contains both HDR10plus and Dovi VSVDBs, we are
incorrectly resetting both dovi and hdr10plus parsed info structs.
As a result, we are only left with the parsed info of the later
VSVDB in the EDID. So, if we have a hdr10plus VSVDB after dovi
VSVDB, dovi info will be reset while parsing for hdr10plus, and
we will not report any dovi capabilities to the frameworks. This
is what is happening on these TVs in question.

Parse HDR10plus and Dovi VSVDBs independently of one another. Do
not reset parsed info of one while parsing for the other.

Since Dovi->HDR10+->Dovi transitions are not defined, send zero
drm/vsif packets while exiting HDR10+ playback if going to dovi
mode next.

Verify:
HDR10plus and Dovi capabilities can be simultaneously and correctly
reported to the frameworks.

Change-Id: I7a344638e5a923c88ef23f0b3a03480c5a564548
Signed-off-by: Zongdong Jiao <zongdong.jiao@amlogic.com>
Signed-off-by: August <mingxin.android@gmail.com>