Video: add pts rollback detect [1/1]
PD#SWPL-23306

Problem:
For the OTT-9111, when application start playing
before dvb stream rollback, the video pts is
bigger than audio because audio has been dropped
the tail of the stream. Video can`t toggle frame
in A-master mode.

Solution:
Force video toggle the video frame if meet the
tail frame of the stream.I made a parameter to
set the threshold between the audio and video pts.
This solution Will influence the stream which has
pts error, need QA to verify that stream on DVB
environment.

Verify:
S905X2

Change-Id: Idc8f050cdb9f5d8c878f3570e2213a5e1f309ca1
Signed-off-by: junliang zhou <junliang.zhou@amlogic.com>
Signed-off-by: August <mingxin.android@gmail.com>